add comment function

+++ models/posts.js
@@ -1,5 +1,27 @@
 const marked = require('marked');
 const Post = require('../lib/mongo').Post;
+const CommentModel = require('./comments');
+
+// post add commentsCount
+Post.plugin('addCommentsCount', {
+  afterFind: function (posts) {
+    return Promise.all(posts.map(function (post) {
+      return CommentModel.getCommentsCount(post._id).then(function (commentsCount) {
+        post.commentsCount = commentsCount;
+        return post;
+      });
+    }));
+  },
+  afterFindOne: function (post) {
+    if (post) {
+      return CommentModel.getCommentsCount(post._id).then(function (count) {
+        post.commentsCount = count;
+        return post;
+      });
+    }
+    return post;
+  }
+});
 
 // convert post content to html
 Post.plugin('contentToHtml', {
@@ -29,6 +51,7 @@ module.exports = {
       .findOne({ _id: postId })
       .populate({ path: 'author', model: 'User' })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec();
   },
@@ -44,6 +67,7 @@ module.exports = {
       .populate({ path: 'author', model: 'User' })
       .sort({ _id: -1 })
       .addCreatedAt()
+      .addCommentsCount()
       .contentToHtml()
       .exec();
   },
@@ -69,7 +93,12 @@ module.exports = {
   },
 
   // delete post by id
-  delPostById: function delPostById (postId) {
-    return Post.deleteOne({ _id: postId }).exec();
+  delPostById: function delPostById (postId, author) {
+    return Post.deleteOne({ author: author, _id: postId }).exec().then(function (res) {
+      // deleted, then delete all comments in the deleted article
+      if (res.result.ok && res.result.n > 0) {
+        return CommentModel.delCommentsByPostId(postId);
+      }
+    });
   }
 };
